refactor: single-source DashboardJsonData in dashboard-data.ts (#965)
dashboard-server.ts redefined DashboardJsonData inline. Any shape
change to the server's JSON response had to be mirrored in two places,
and any drift was silent — the CLI emitted one shape while the server's
declared type told callers it emitted another.

Move the interface to dashboard-data.ts (next to DashboardStats and
DashboardFetchResult) and import it from dashboard-server.ts. Removes
five now-unused type imports from dashboard-server.ts (FetchedPR,
ShelvedPRRef, ParseIssueListOutput was never directly used, etc.).

Short-term fix only — this closes the drift risk for DashboardJsonData,
but the #965 issue also mentioned a broader runtime --json contract
enforcement concern. That's kept open as a follow-up design task; it
needs a Zod-schemas-at-the-boundary vs. golden-file-tests decision that
shouldn't be rolled into this refactor.
